requirements.txt
# requirements.txt
# Core dependencies for Hotel Booking API

# LangChain and LLM dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Additional utilities
python-dotenv>=1.0.0
httpx>=0.25.0

# Optional: For enhanced development
# black>=23.0.0          # Code formatting
# pytest>=7.4.0          # Testing
# pytest-asyncio>=0.21.0 # Async testing